Sync CIRCABC OSS with enterprise develop-oss branch - Aug 2026

+++ circabc-backend/source/java/eu/cec/digit/circabc/repo/web/scripts/bean/GroupLogosDelete.java
@@ -56,6 +56,11 @@ public class GroupLogosDelete extends CircabcDeclarativeWebScript {
           "Cannot delete a logo, not enough permissions"
         );
       }
+      if (!this.groupLockApi.canWriteOrAdmin(groupId)) {
+        throw new ReadOnlyAccessException(
+          "Interest group is in read-only mode"
+        );
+      }
 
       if (groupId != null) {
         this.groupsApi.deleteLogo(groupId, logoId);
@@ -64,11 +69,27 @@ public class GroupLogosDelete extends CircabcDeclarativeWebScript {
       status.setCode(HttpServletResponse.SC_FORBIDDEN);
       status.setMessage("Access denied");
       status.setRedirect(true);
+      if (logger.isErrorEnabled()) {
+        logger.error("Access denied", ade);
+      }
+      return null;
+    } catch (ReadOnlyAccessException roe) {
+      status.setCode(HttpServletResponse.SC_FORBIDDEN);
+      status.setMessage(roe.getMessage());
+      status.setRedirect(true);
+      if (logger.isWarnEnabled()) {
+        logger.warn(
+          "Read-only mode prevented logo delete: " + roe.getMessage()
+        );
+      }
       return null;
     } catch (InvalidNodeRefException | CustomizationException inre) {
       status.setCode(HttpServletResponse.SC_BAD_REQUEST);
       status.setMessage("Bad request");
       status.setRedirect(true);
+      if (logger.isErrorEnabled()) {
+        logger.error("Bad request", inre);
+      }
       return null;
     } finally {
       MLPropertyInterceptor.setMLAware(mlAware);
